V1.0.5

+++ database/migrations/2023_04_07_005723_create_rol_user_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('rol_user', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('rol_id');
-            $table->unsignedBigInteger('rol_user');
+            $table->Integer('rol_id');
+            $table->Integer('rol_user');
             $table->foreign('rol_id')->references('id')->on('roles');
             $table->foreign('user_id')->references('id')->on('users');
             $table->timestamps();
